requirements.txt

# Django
Django==2.2.7

# Database
psycopg2===2.7.3.2



# Rest Framework
djangorestframework==3.10.3
django-filter==2.2.0

